+++ receiver/swok8sdiscovery/ticker.go
@@ -1,0 +1,42 @@
+// Copyright 2025 SolarWinds Worldwide, LLC. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package swok8sdiscovery
+
+import (
+	"context"
+	"time"
+)
+
+// Start ticking immediately.
+// Ref: https://stackoverflow.com/questions/32705582/how-to-get-time-tick-to-tick-immediately
+func newTicker(ctx context.Context, repeat time.Duration) *time.Ticker {
+	ticker := time.NewTicker(repeat)
+	oc := ticker.C
+	nc := make(chan time.Time, 1)
+	go func() {
+		nc <- time.Now()
+		for {
+			select {
+			case tm := <-oc:
+				nc <- tm
+			case <-ctx.Done():
+				return
+			}
+		}
+	}()
+
+	ticker.C = nc
+	return ticker
+}