+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## vNext
 - Add `mqttreceiver`
 - `swohostmetricsreceiver` utilizes OTEL generated code instead of original internal implementation. Relevant code cleanup.
+- Added receiver `swok8sdiscovery` to publish known database entities and relationships
 
 ## v0.131.9
 - Updates golang to 1.25.1